Poprawa wpisywania grafu

+++ Projekt/ConvexHull.h
@@ -7,7 +7,7 @@
 class ConvexHull 
 {
 private:
-    static int crossProduct(Vertex p, Vertex q, Vertex r);
+    static float crossProduct(Vertex p, Vertex q, Vertex r);
     static float distance(Vertex p, Vertex q);
 public:
     static std::vector<Vertex> calculateConvexHull(std::vector<Vertex> v); // znajdowanie otoczki
